refactor: break up 7 largest view files into focused sub-views (#603)
Extract self-contained types and >100-line blocks from oversized files:

- LauncherCollectionView (1056→206): extract LauncherMappingEditor
- MapperViewModel+ConflictResolution (968→53): extract CaptureManagement + RuleManagement
- LiveKeyboardOverlayView+Header (982→757): extract LayerPicker popover
- KeyboardSelectionGridView (918→607): extract KeyboardIllustrationCard + LayoutCategorySection
- OverlayKeyboardView (913→702): extract FloatingKeymapLabel
- ChordGroupsModalView (810→534): extract ChordEditorDialog
- HomeRowModsCollectionView (805→311): extract Configuration + Popovers

3 files skipped (RulesSummaryView, BaseKeycap, ContextHUDController) —
no individual section exceeded 100 lines.

Pure structural refactor, no behavior changes.
