+++ Sources/KeyPathAppKit/UI/Overlay/KeyboardSelectionGridView.swift
@@ -394,318 +394,8 @@ struct KeyboardSelectionGridView: View {
     }
 }
 
-/// A section with a category header and its layouts
-private struct LayoutCategorySection: View {
-    let category: LayoutCategory
-    let layouts: [PhysicalLayout]
-    @Binding var selectedLayoutId: String
-    let isDark: Bool
-    @Binding var showImportSheet: Bool
-    @Binding var refreshTrigger: UUID
-    var onToast: ((String, KanataViewModel.ToastType) -> Void)?
-
-    private let columns = [
-        GridItem(.flexible(), spacing: 12)
-    ]
-
-    /// Whether to show section header (US Standard has no header)
-    private var showHeader: Bool {
-        category != .usStandard
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            // Section header (hidden for US Standard)
-            if showHeader {
-                HStack {
-                    Text(category.displayName)
-                        .font(.system(size: 12, weight: .semibold))
-                        .foregroundStyle(.primary.opacity(0.7))
-                        .textCase(.uppercase)
-                        .tracking(0.8)
-
-                    Spacer()
-
-                    // Import button for Custom section
-                    if category == .custom {
-                        Button {
-                            showImportSheet = true
-                        } label: {
-                            HStack(spacing: 4) {
-                                Image(systemName: "plus.circle.fill")
-                                    .font(.system(size: 10))
-                                Text("Import")
-                                    .font(.system(size: 10, weight: .medium))
-                            }
-                            .foregroundStyle(.secondary)
-                        }
-                        .buttonStyle(.plain)
-                        .accessibilityIdentifier("import-custom-layout-button")
-                    }
-                }
-                .padding(.leading, 4)
-                .padding(.trailing, 4)
-                .padding(.top, 8)
-            }
-
-            // Layout cards in grid
-            LazyVGrid(columns: columns, spacing: 16) {
-                ForEach(layouts) { layout in
-                    KeyboardIllustrationCard(
-                        layout: layout,
-                        isSelected: selectedLayoutId == layout.id,
-                        isDark: isDark,
-                        isCustom: layout.id.hasPrefix("custom-")
-                    ) {
-                        // No animation on selection - keeps drawer stable
-                        selectedLayoutId = layout.id
-                    } onDelete: {
-                        if layout.id.hasPrefix("custom-") {
-                            Task {
-                                let layoutId = String(layout.id.dropFirst(7))
-                                await QMKImportService.shared.deleteCustomLayout(layoutId: layoutId)
-                                // If deleted layout was selected, reset to default
-                                if selectedLayoutId == layout.id {
-                                    selectedLayoutId = LayoutPreferences.defaultLayoutId
-                                }
-                                // Trigger refresh to update UI
-                                refreshTrigger = UUID()
-                            }
-                        }
-                    } onRefreshKeymap: {
-                        if layout.id.hasPrefix("custom-") {
-                            Task {
-                                let layoutId = String(layout.id.dropFirst(7))
-                                let result = await QMKImportService.shared.refreshKeymap(layoutId: layoutId)
-                                await MainActor.run {
-                                    switch result {
-                                    case let .success(tokenCount):
-                                        refreshTrigger = UUID()
-                                        onToast?("Keymap refreshed — \(tokenCount) keys mapped", .success)
-                                    case let .failure(message):
-                                        onToast?(message, .warning)
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    .id("layout-\(layout.id)")
-                }
-
-                // Import button card for Custom section (if no layouts yet)
-                if category == .custom, layouts.isEmpty {
-                    ImportLayoutCard(isDark: isDark) {
-                        showImportSheet = true
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Individual keyboard illustration card with image and label
-private struct KeyboardIllustrationCard: View {
-    let layout: PhysicalLayout
-    let isSelected: Bool
-    let isDark: Bool
-    let isCustom: Bool
-    let onSelect: () -> Void
-    let onDelete: (() -> Void)?
-    let onRefreshKeymap: (() -> Void)?
-
-    @State private var isHovering = false
-    @State private var showDeleteConfirmation = false
-
-    /// Image size - fairly big as requested
-    private let imageHeight: CGFloat = 120
-
-    init(
-        layout: PhysicalLayout,
-        isSelected: Bool,
-        isDark: Bool,
-        isCustom: Bool = false,
-        onSelect: @escaping () -> Void,
-        onDelete: (() -> Void)? = nil,
-        onRefreshKeymap: (() -> Void)? = nil
-    ) {
-        self.layout = layout
-        self.isSelected = isSelected
-        self.isDark = isDark
-        self.isCustom = isCustom
-        self.onSelect = onSelect
-        self.onDelete = onDelete
-        self.onRefreshKeymap = onRefreshKeymap
-    }
-
-    var body: some View {
-        Button(action: {
-            // Haptic feedback
-            let generator = NSHapticFeedbackManager.defaultPerformer
-            generator.perform(.alignment, performanceTime: .default)
-
-            onSelect()
-        }) {
-            VStack(spacing: 8) {
-                // Keyboard illustration - fixed height, horizontally centered
-                ZStack {
-                    RoundedRectangle(cornerRadius: 12, style: .continuous)
-                        .fill(cardBackground)
-                        .shadow(
-                            color: shadowColor,
-                            radius: shadowRadius,
-                            x: 0,
-                            y: shadowY
-                        )
-
-                    // Selection ring overlay
-                    if isSelected {
-                        RoundedRectangle(cornerRadius: 12, style: .continuous)
-                            .strokeBorder(Color.accentColor.opacity(0.6), lineWidth: 2)
-                    }
-
-                    keyboardImage
-                        .frame(maxHeight: imageHeight)
-                        .padding(12)
-                }
-                .frame(height: imageHeight + 24)
-
-                // Label - centered below image, supports multi-line for long names
-                Text(layout.name)
-                    .font(.system(size: 11, weight: isSelected ? .semibold : .medium))
-                    .foregroundStyle(isSelected ? .primary : .secondary)
-                    .multilineTextAlignment(.center)
-                    .lineLimit(nil)
-                    .fixedSize(horizontal: false, vertical: true)
-                    .frame(maxWidth: .infinity)
-            }
-            .frame(maxWidth: .infinity)
-        }
-        .buttonStyle(.plain)
-        .accessibilityIdentifier("overlay-keyboard-layout-button-\(layout.id)")
-        .accessibilityLabel("Select keyboard layout \(layout.name)")
-        .onHover { hovering in
-            // Animate only hover state, not scale - avoids layout shifts
-            withAnimation(.easeInOut(duration: 0.15)) {
-                isHovering = hovering
-            }
-        }
-        // No scale "pop" animation on selection - keeps drawer stable
-        // Selection ring and shadow changes provide sufficient visual feedback
-        .contextMenu {
-            if isCustom {
-                if onRefreshKeymap != nil {
-                    Button {
-                        onRefreshKeymap?()
-                    } label: {
-                        Label("Refresh Keymap", systemImage: "arrow.clockwise")
-                    }
-                    .accessibilityIdentifier("refresh-keymap-button")
-                }
-
-                if onDelete != nil {
-                    Button(role: .destructive) {
-                        showDeleteConfirmation = true
-                    } label: {
-                        Label("Delete Layout", systemImage: "trash")
-                    }
-                    .accessibilityIdentifier("delete-custom-layout-button")
-                }
-            }
-        }
-        .confirmationDialog(
-            "Delete Layout",
-            isPresented: $showDeleteConfirmation,
-            titleVisibility: .visible
-        ) {
-            Button("Delete", role: .destructive) {
-                onDelete?()
-            }
-            .accessibilityIdentifier("overlay-keyboard-layout-delete-confirm")
-            Button("Cancel", role: .cancel) {}
-                .accessibilityIdentifier("overlay-keyboard-layout-delete-cancel")
-        } message: {
-            Text("Are you sure you want to delete \"\(layout.name)\"? This action cannot be undone.")
-        }
-    }
-
-    // MARK: - Computed Animation Properties
-
-    private var shadowColor: Color {
-        if isSelected {
-            Color.accentColor.opacity(0.4)
-        } else if isHovering {
-            Color.black.opacity(0.15)
-        } else {
-            Color.black.opacity(0.08)
-        }
-    }
-
-    private var shadowRadius: CGFloat {
-        if isSelected {
-            10
-        } else if isHovering {
-            6
-        } else {
-            4
-        }
-    }
-
-    private var shadowY: CGFloat {
-        if isSelected {
-            5
-        } else if isHovering {
-            3
-        } else {
-            2
-        }
-    }
-
-    @ViewBuilder
-    private var keyboardImage: some View {
-        // Images are at bundle root (not in subdirectory) due to .process() flattening
-        // Same pattern as SVG loading in LiveKeyboardOverlayView
-        let imageURL = KeyPathAppKitResources.url(
-            forResource: layout.id,
-            withExtension: "png"
-        ) ?? Bundle.main.url(
-            forResource: layout.id,
-            withExtension: "png"
-        )
-
-        if let url = imageURL,
-           let image = NSImage(contentsOf: url)
-        {
-            Image(nsImage: image)
-                .resizable()
-                .interpolation(.high)
-                .aspectRatio(contentMode: .fit)
-        } else {
-            // Fallback to SF Symbol
-            Image(systemName: "keyboard")
-                .font(.system(size: 48))
-                .foregroundStyle(.secondary)
-        }
-    }
-
-    private var cardBackground: some ShapeStyle {
-        if isSelected {
-            AnyShapeStyle(
-                Color.accentColor.opacity(isDark ? 0.25 : 0.18)
-            )
-        } else if isHovering {
-            AnyShapeStyle(
-                Color.white.opacity(isDark ? 0.12 : 0.10)
-            )
-        } else {
-            AnyShapeStyle(
-                Color.white.opacity(isDark ? 0.06 : 0.05)
-            )
-        }
-    }
-}
-
 /// Card for importing a custom layout (shown when Custom section is empty)
-private struct ImportLayoutCard: View {
+struct ImportLayoutCard: View {
     let isDark: Bool
     let onImport: () -> Void
