+++ Sources/KeyPathAppKit/UI/Rules/HomeRowModsCollectionView.swift
@@ -13,28 +13,29 @@ struct HomeRowModsCollectionView: View {
     var holdTimingBinding: Binding<Double>?
     var holdTimingValue: Int = 180
     var holdTimingDefault: Int = 180
-    var holdTimingRange: ClosedRange<Double> = 120...300
+    var holdTimingRange: ClosedRange<Double> = 120 ... 300
     var holdTimingStep: Double = 20
     var onResetHoldTiming: (() -> Void)?
     var isCollectionEnabled: Bool = true
 
     @AppStorage(KeymapPreferences.keymapIdKey) private var selectedKeymapId: String = LogicalKeymap.defaultId
-    @State private var showingCustomizeWindow = false
-    @State private var selectedKey: String?
-    @State private var showingNewLayerSheet = false
-    @State private var newLayerName = ""
-    @State private var locallyCreatedLayers: Set<String> = []
-    @State private var hoveredHoldBehavior: HomeRowHoldMode?
-    @State private var hoveredLayerToggleMode: LayerToggleMode?
-    @State private var timingPreviewPhase: TimingPreviewPhase = .idle
-    @State private var timingPreviewTask: Task<Void, Never>?
+    @State var showingCustomizeWindow = false
+    @State var selectedKey: String?
+    @State var showingNewLayerSheet = false
+    @State var newLayerName = ""
+    @State var locallyCreatedLayers: Set<String> = []
+    @State var hoveredHoldBehavior: HomeRowHoldMode?
+    @State var hoveredLayerToggleMode: LayerToggleMode?
+    @State var timingPreviewPhase: TimingPreviewPhase = .idle
+    @State var timingPreviewTask: Task<Void, Never>?
 
     enum TimingPreviewPhase: Equatable {
         case idle
         case pressing
         case modifier
     }
-    @State private var showingHelp = false
+
+    @State var showingHelp = false
 
     private var activeKeymap: LogicalKeymap {
         .resolve(id: selectedKeymapId)
@@ -117,53 +118,7 @@ struct HomeRowModsCollectionView: View {
         }
     }
 
-    // MARK: - Customize Section
-
-    private var customizeWindowContent: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            HStack {
-                Text("Home Row Preferences")
-                    .font(.title3.weight(.semibold))
-
-                Button { showingHelp = true } label: {
-                    Image(systemName: "questionmark.circle")
-                        .foregroundColor(.secondary)
-                }
-                .buttonStyle(.plain)
-                .accessibilityIdentifier("home-row-mods-prefs-help-button")
-                .accessibilityLabel("Home row mods help")
-
-                Spacer()
-
-                Button("Done") {
-                    showingCustomizeWindow = false
-                }
-                .keyboardShortcut(.defaultAction)
-                .accessibilityIdentifier("home-row-mods-done-button")
-            }
-            .padding(.horizontal, 20)
-            .padding(.top, 16)
-            .padding(.bottom, 10)
-
-            Divider()
-
-            ScrollView {
-                customizeSection
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(20)
-            }
-        }
-        .frame(minWidth: 560, minHeight: 520)
-        .settingsBackground()
-        .sheet(isPresented: $showingNewLayerSheet) {
-            newLayerSheet
-        }
-        .sheet(isPresented: $showingHelp) {
-            MarkdownHelpSheet(resource: "home-row-mods", title: "Home Row Mods")
-        }
-    }
-
-    private var usesTopRowKeys: Bool {
+    var usesTopRowKeys: Bool {
         config.enabledKeys.contains(where: { HomeRowModsConfig.vallackTopRowKeys.contains($0) })
     }
 
@@ -201,295 +156,7 @@ struct HomeRowModsCollectionView: View {
         )
     }
 
-    private var customizeSection: some View {
-        VStack(alignment: .leading, spacing: 20) {
-            Text(usesTopRowKeys
-                ? "These settings control what your top-row keys do when you hold them. Tap still types letters."
-                : "These settings control what your home-row keys do when you hold them. Tap still types letters.")
-                .font(.subheadline)
-                .foregroundStyle(.secondary)
-
-            VStack(alignment: .leading, spacing: 10) {
-                Text("Hold Behavior")
-                    .font(.headline)
-                    .foregroundStyle(.secondary)
-
-                HStack(spacing: 10) {
-                    holdBehaviorOptionCard(
-                        mode: .modifiers,
-                        icon: "command",
-                        title: "Modifiers",
-                        subtitle: "Recommended"
-                    )
-
-                    holdBehaviorOptionCard(
-                        mode: .layers,
-                        icon: "square.stack.3d.up",
-                        title: "Layers",
-                        subtitle: "Advanced"
-                    )
-                }
-                .accessibilityIdentifier("home-row-mods-hold-mode-picker")
-                .accessibilityLabel("Hold action mode")
-
-                Text(holdBehaviorExplanationText)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-
-            if config.holdMode == .layers {
-                Divider()
-
-                VStack(alignment: .leading, spacing: 10) {
-                    Text("Layer Activation")
-                        .font(.headline)
-                        .foregroundStyle(.secondary)
-
-                    HStack(spacing: 10) {
-                        SettingsOptionCard(
-                            icon: "hand.raised",
-                            title: "While Held",
-                            subtitle: "Active only while holding",
-                            isSelected: config.layerToggleMode == .whileHeld,
-                            onHoverChanged: { hovering in
-                                hoveredLayerToggleMode = hovering ? .whileHeld : nil
-                            }
-                        ) {
-                            config.layerToggleMode = .whileHeld
-                            updateConfig()
-                        }
-                        .accessibilityIdentifier("home-row-mods-layer-toggle-while-held")
-
-                        SettingsOptionCard(
-                            icon: "switch.2",
-                            title: "Toggle",
-                            subtitle: "Press once to stay on",
-                            isSelected: config.layerToggleMode == .toggle,
-                            onHoverChanged: { hovering in
-                                hoveredLayerToggleMode = hovering ? .toggle : nil
-                            }
-                        ) {
-                            config.layerToggleMode = .toggle
-                            updateConfig()
-                        }
-                        .accessibilityIdentifier("home-row-mods-layer-toggle-mode")
-                    }
-                    .accessibilityElement(children: .contain)
-                    .accessibilityIdentifier("home-row-mods-layer-toggle-mode-picker")
-                    .accessibilityLabel("Layer activation mode")
-
-                    Text(layerActivationExplanationText)
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                }
-            }
-
-            Divider()
-
-            HomeRowTimingSection(
-                config: $config,
-                showsHrmInsights: true,
-                onConfigChanged: onConfigChanged
-            )
-            .frame(maxWidth: 500)
-        }
-    }
-
-    // MARK: - Assignment Popovers
-
-    private var modifierOptions: [(kind: String, label: String, symbol: String)] {
-        [
-            ("met", "Command", "⌘"),
-            ("alt", "Option", "⌥"),
-            ("ctl", "Control", "⌃"),
-            ("sft", "Shift", "⇧")
-        ]
-    }
-
-    private func modifierPopoverContent(for key: String) -> some View {
-        VStack(spacing: 0) {
-            ForEach(modifierOptions.indices, id: \.self) { index in
-                let option = modifierOptions[index]
-                Button {
-                    config.modifierAssignments[key] = sidedModifierAssignment(for: key, kind: option.kind)
-                    updateConfig()
-                    selectedKey = nil
-                } label: {
-                    HStack(spacing: 10) {
-                        Text(option.symbol)
-                            .font(.body.weight(.semibold))
-                            .frame(width: 20)
-                            .foregroundStyle(isModifierOptionSelected(option.kind, for: key) ? Color.accentColor : .secondary)
-                        Text(option.label)
-                            .font(.body)
-                        Spacer()
-                        if isModifierOptionSelected(option.kind, for: key) {
-                            Image(systemName: "checkmark")
-                                .font(.body.weight(.semibold))
-                                .foregroundStyle(Color.accentColor)
-                        }
-                    }
-                    .foregroundStyle(.primary)
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 10)
-                    .contentShape(Rectangle())
-                }
-                .buttonStyle(LayerPickerItemButtonStyle())
-                .focusable(false)
-                .accessibilityIdentifier("home-row-mods-modifier-option-\(option.kind)")
-                .accessibilityLabel("Set \(displayLabel(forCanonicalKey: key)) to \(option.symbol) \(option.label)")
-
-                if index < modifierOptions.count - 1 {
-                    PopoverListDivider()
-                }
-            }
-
-            PopoverListDivider()
-
-            disableKeyButton(for: key)
-        }
-        .padding(.vertical, 6)
-        .frame(minWidth: 210)
-        .pickerPopoverChrome()
-    }
-
-    private func layerPopoverContent(for key: String) -> some View {
-        VStack(spacing: 0) {
-            ForEach(layerOptions.indices, id: \.self) { index in
-                let option = layerOptions[index]
-                Button {
-                    config.layerAssignments[key] = option.key
-                    updateConfig()
-                    selectedKey = nil
-                } label: {
-                    HStack(spacing: 10) {
-                        Image(systemName: option.icon)
-                            .font(.body)
-                            .frame(width: 20)
-                            .foregroundStyle(config.layerAssignments[key] == option.key ? Color.accentColor : .secondary)
-                        Text(option.label)
-                            .font(.body)
-                        Spacer()
-                        if config.layerAssignments[key] == option.key {
-                            Image(systemName: "checkmark")
-                                .font(.body.weight(.semibold))
-                                .foregroundStyle(Color.accentColor)
-                        }
-                    }
-                    .foregroundStyle(.primary)
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 10)
-                    .contentShape(Rectangle())
-                }
-                .buttonStyle(LayerPickerItemButtonStyle())
-                .focusable(false)
-                .accessibilityIdentifier("home-row-mods-layer-option-\(option.key)")
-                .accessibilityLabel("Set \(displayLabel(forCanonicalKey: key)) to \(option.label) layer")
-
-                if index < layerOptions.count - 1 {
-                    PopoverListDivider()
-                }
-            }
-
-            PopoverListDivider()
-
-            disableKeyButton(for: key)
-
-            PopoverListDivider()
-
-            Button {
-                showingNewLayerSheet = true
-            } label: {
-                HStack(spacing: 10) {
-                    Image(systemName: "plus")
-                        .font(.body)
-                        .frame(width: 20)
-                        .foregroundStyle(.secondary)
-                    Text("New Layer...")
-                        .font(.body)
-                    Spacer()
-                }
-                .foregroundStyle(.primary)
-                .padding(.horizontal, 12)
-                .padding(.vertical, 10)
-                .contentShape(Rectangle())
-            }
-            .buttonStyle(LayerPickerItemButtonStyle())
-            .focusable(false)
-            .accessibilityIdentifier("home-row-mods-layer-option-new")
-            .accessibilityLabel("Create a new layer")
-        }
-        .padding(.vertical, 6)
-        .frame(minWidth: 210)
-        .pickerPopoverChrome()
-    }
-
-    private func enableKeyPopoverContent(for key: String) -> some View {
-        VStack(spacing: 0) {
-            Button {
-                config.enabledKeys.insert(key)
-                // Restore default assignment for this key's position
-                if config.holdMode == .modifiers {
-                    config.modifierAssignments[key] = HomeRowModsConfig.cagsMacDefault[key]
-                } else {
-                    config.layerAssignments[key] = HomeRowModsConfig.defaultLayerAssignments[key]
-                }
-                config.keySelection = .custom
-                updateConfig()
-                selectedKey = nil
-            } label: {
-                HStack(spacing: 10) {
-                    Image(systemName: "checkmark")
-                        .font(.body.weight(.semibold))
-                        .frame(width: 20)
-                        .foregroundStyle(Color.accentColor)
-                    Text("Enable Key")
-                        .font(.body)
-                    Spacer()
-                }
-                .foregroundStyle(.primary)
-                .padding(.horizontal, 12)
-                .padding(.vertical, 10)
-                .contentShape(Rectangle())
-            }
-            .buttonStyle(LayerPickerItemButtonStyle())
-            .focusable(false)
-            .accessibilityIdentifier("home-row-mods-enable-key-\(key)")
-            .accessibilityLabel("Enable \(displayLabel(forCanonicalKey: key))")
-        }
-        .padding(.vertical, 6)
-        .frame(minWidth: 210)
-        .pickerPopoverChrome()
-    }
-
-    private func disableKeyButton(for key: String) -> some View {
-        Button {
-            config.enabledKeys.remove(key)
-            config.keySelection = .custom
-            updateConfig()
-            selectedKey = nil
-        } label: {
-            HStack(spacing: 10) {
-                Image(systemName: "nosign")
-                    .font(.body)
-                    .frame(width: 20)
-                    .foregroundStyle(.secondary)
-                Text("Disable Key")
-                    .font(.body)
-                Spacer()
-            }
-            .foregroundStyle(.primary)
-            .padding(.horizontal, 12)
-            .padding(.vertical, 10)
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(LayerPickerItemButtonStyle())
-        .focusable(false)
-        .accessibilityIdentifier("home-row-mods-disable-key-\(key)")
-        .accessibilityLabel("Disable \(displayLabel(forCanonicalKey: key))")
-    }
-
-    private var layerOptions: [(key: String, label: String, icon: String)] {
+    var layerOptions: [(key: String, label: String, icon: String)] {
         let options = availableLayerNames.map { layer in
             (key: layer, label: LayerInfo.displayName(for: layer), icon: LayerInfo.iconName(for: layer))
         }
@@ -502,15 +169,15 @@ struct HomeRowModsCollectionView: View {
 
     // MARK: - Helpers
 
-    private func updateConfig() {
+    func updateConfig() {
         onConfigChanged(config)
     }
 
-    private var activeHoldAssignments: [String: String] {
+    var activeHoldAssignments: [String: String] {
         config.holdMode == .modifiers ? config.modifierAssignments : config.layerAssignments
     }
 
-    private func displayLabel(forCanonicalKey key: String) -> String {
+    func displayLabel(forCanonicalKey key: String) -> String {
         guard let keyCode = LogicalKeymap.keyCode(forQwertyLabel: key),
               let label = activeKeymap.label(for: keyCode, includeExtraKeys: false)
         else {
@@ -526,7 +193,7 @@ struct HomeRowModsCollectionView: View {
         return "Click a key to choose its modifier."
     }
 
-    private var knownAvailableLayerNames: Set<String> {
+    var knownAvailableLayerNames: Set<String> {
         Set(
             availableLayers
                 .map { $0.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() }
@@ -538,17 +205,17 @@ struct HomeRowModsCollectionView: View {
         )
     }
 
-    private var availableLayerNames: [String] {
+    var availableLayerNames: [String] {
         let names = Array(knownAvailableLayerNames).sorted()
         assert(Set(names.map { $0.lowercased() }) == knownAvailableLayerNames, "Layer names should be normalized.")
         return names
     }
 
-    private var recommendedLayerNames: [String] {
+    var recommendedLayerNames: [String] {
         ["fun", "num", "sym", "nav"]
     }
 
-    private var recommendedLayerAssignments: [String: String] {
+    var recommendedLayerAssignments: [String: String] {
         let left = recommendedLayerNames
         return [
             "a": left[0], "s": left[1], "d": left[2], "f": left[3],
@@ -582,13 +249,6 @@ struct HomeRowModsCollectionView: View {
     private func normalizeLegacyLayerModeIfNeeded() {
         guard config.holdMode == .layers, config.hasUserSelectedHoldMode == false else { return }
         config.holdMode = .modifiers
-        selectedKey = nil
-        updateConfig()
-    }
-
-    private func applyHoldMode(_ mode: HomeRowHoldMode) {
-        config.holdMode = mode
-        config.hasUserSelectedHoldMode = true
         selectedKey = nil
         updateConfig()
     }
@@ -647,159 +307,5 @@ struct HomeRowModsCollectionView: View {
         config.holdMode = baseline.holdMode
         onResetHoldTiming?()
         updateConfig()
-    }
-
-    @ViewBuilder
-    private func holdBehaviorOptionCard(
-        mode: HomeRowHoldMode,
-        icon: String,
-        title: String,
-        subtitle: String
-    ) -> some View {
-        let isSelected = config.holdMode == mode
-        let isHovered = hoveredHoldBehavior == mode
-
-        Button {
-            if mode == .layers {
-                Task {
-                    let ids = [
-                        RuleCollectionIdentifier.vimNavigation,
-                        RuleCollectionIdentifier.symbolLayer,
-                        RuleCollectionIdentifier.numpadLayer,
-                        RuleCollectionIdentifier.funLayer
-                    ]
-                    await onEnableLayerCollections?(ids)
-                    config.layerAssignments = recommendedLayerAssignments
-                    applyHoldMode(.layers)
-                }
-                return
-            }
-            applyHoldMode(mode)
-        } label: {
-            HStack(spacing: 10) {
-                Image(systemName: icon)
-                    .font(.headline)
-                    .foregroundStyle(isSelected ? Color.accentColor : .secondary)
-                    .frame(width: 24)
-
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(title)
-                        .font(.subheadline.weight(isSelected ? .semibold : .regular))
-                        .foregroundStyle(isSelected ? .primary : .secondary)
-                    Text(subtitle)
-                        .font(.caption2)
-                        .foregroundStyle(.tertiary)
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .padding(.vertical, 10)
-            .padding(.horizontal, 12)
-            .contentShape(Rectangle())
-            .background(
-                RoundedRectangle(cornerRadius: 8, style: .continuous)
-                    .fill(isSelected ? Color.accentColor.opacity(0.08) : .clear)
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 8, style: .continuous)
-                    .strokeBorder(
-                        isSelected
-                            ? Color.accentColor.opacity(0.35)
-                            : (isHovered ? Color.accentColor.opacity(0.35) : Color.primary.opacity(0.08)),
-                        lineWidth: isSelected ? 1.5 : (isHovered ? 1.5 : 0.5)
-                    )
-            )
-        }
-        .buttonStyle(.plain)
-        .focusable(false)
-        .onHover { hovering in
-            hoveredHoldBehavior = hovering ? mode : (hoveredHoldBehavior == mode ? nil : hoveredHoldBehavior)
-        }
-        .accessibilityIdentifier(
-            mode == .modifiers ? "home-row-mods-hold-mode-modifiers" : "home-row-mods-hold-mode-layers"
-        )
-    }
-
-    private var holdBehaviorExplanationText: String {
-        let keyDesc = usesTopRowKeys ? "a top-row key" : "a home-row key"
-        switch hoveredHoldBehavior ?? config.holdMode {
-        case .modifiers:
-            return "Best for shortcuts and general app use. Hold \(keyDesc) to get Cmd/Opt/Ctrl/Shift."
-        case .layers:
-            return "Best for advanced layouts. Hold \(keyDesc) to temporarily access another key layer."
-        }
-    }
-
-    private var layerActivationExplanationText: String {
-        switch hoveredLayerToggleMode ?? config.layerToggleMode {
-        case .whileHeld:
-            "Momentary mode: layer turns off when you release the key."
-        case .toggle:
-            "Latch mode: layer stays active until you switch it off."
-        }
-    }
-
-    private func layerPresetSelection(from assignments: [String: String]) -> HomeRowLayerPreset {
-        if assignments == recommendedLayerAssignments { return .default }
-        return .custom
-    }
-
-    private var newLayerSheet: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text("Create New Layer")
-                .font(.title3.weight(.semibold))
-
-            TextField("Layer name", text: $newLayerName)
-                .textFieldStyle(.roundedBorder)
-                .accessibilityIdentifier("home-row-mods-new-layer-name")
-
-            HStack {
-                Spacer()
-                Button("Cancel") {
-                    newLayerName = ""
-                    showingNewLayerSheet = false
-                }
-                .keyboardShortcut(.cancelAction)
-                .accessibilityIdentifier("home-row-mods-new-layer-cancel-button")
-
-                Button("Create") {
-                    Task {
-                        let raw = newLayerName.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-                        guard !raw.isEmpty else { return }
-                        await onEnsureLayersExist([raw])
-                        locallyCreatedLayers.insert(raw)
-                        if let selectedKey {
-                            config.layerAssignments[selectedKey] = raw
-                            updateConfig()
-                        }
-                        newLayerName = ""
-                        showingNewLayerSheet = false
-                    }
-                }
-                .buttonStyle(.borderedProminent)
-                .keyboardShortcut(.defaultAction)
-                .disabled(newLayerName.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty)
-                .accessibilityIdentifier("home-row-mods-new-layer-create-button")
-            }
-        }
-        .padding(20)
-        .frame(width: 360)
-    }
-
-    private func isModifierOptionSelected(_ kind: String, for key: String) -> Bool {
-        guard let assignment = config.modifierAssignments[key] else { return false }
-        return canonicalModifierKind(from: assignment) == kind
-    }
-
-    private func sidedModifierAssignment(for key: String, kind: String) -> String {
-        let isRightHand = HomeRowModsConfig.rightHandKeys.contains(key)
-        return (isRightHand ? "r" : "l") + kind
-    }
-
-    private func canonicalModifierKind(from assignment: String) -> String {
-        guard assignment.count == 4 else { return assignment }
-        if assignment.hasPrefix("l") || assignment.hasPrefix("r") {
-            return String(assignment.dropFirst())
-        }
-        return assignment
     }
 }
